fix pyright and some ruff issues in src/lmnr/cli
