+++ src/lmnr/cli/dev.py
@@ -8,14 +8,14 @@ import sys
 from argparse import Namespace
 
 
-async def run_dev(args: Namespace) -> None:
+async def run_dev(_args: Namespace) -> None:
     """Show deprecation message for dev command."""
     print(
-        "\033[33m\n"  # Yellow text
-        "The 'lmnr dev' command has been moved to a separate package.\n\n"
-        "Please use:\n"
-        "  npx @lmnr-ai/cli@latest dev [file] [options]\n\n"
-        "See https://docs.lmnr.ai/cli/dev for more information.\n"
+        "\033[33m\n" +  # Yellow text
+        "The 'lmnr dev' command has been moved to a separate package.\n\n" +
+        "Please use:\n" +
+        "  npx @lmnr-ai/cli@latest dev [file] [options]\n\n" +
+        "See https://docs.lmnr.ai/cli/dev for more information.\n" +
         "\033[0m",  # Reset color
         file=sys.stderr,
     )
