--- src/lmnr/sdk/eval_control.py
@@ -1,5 +1,0 @@
-from contextvars import ContextVar
-
-
-PREPARE_ONLY: ContextVar[bool] = ContextVar("__lmnr_prepare_only", default=False)
-EVALUATION_INSTANCES = ContextVar("__lmnr_evaluation_instances")
