+++ src/lmnr/sdk/evaluations/consts.py
@@ -1,0 +1,2 @@
+DEFAULT_BATCH_SIZE = 5
+MAX_EXPORT_BATCH_SIZE = 64
